requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
colorama==0.4.6
qrcode==7.4.2
Werkzeug==3.0.1
apscheduler==3.10.4
pytz==2024.1
WeasyPrint==60.2
pdfkit==1.0.0
celery==5.3.4
redis==5.0.1
flower==2.0.1
psutil==5.9.8
matplotlib>=3.8.0
seaborn==0.12.2
pyotp==2.9.0
alembic==1.13.1
Flask-Migrate==4.0.5
pandas==2.2.0
openpyxl==3.1.2
reportlab==4.0.7
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
gunicorn==21.2.0